build: solve confilcts
