--- src/main/java/com/dottree/nonogrammers/domain/ResponseModel.java
@@ -1,19 +1,0 @@
-package com.dottree.nonogrammers.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Setter
-@Getter
-@ToString
-public class ResponseModel {
-    private Integer statusCode;
-    private String title;
-    private String data;
-    private String message;
-    private HashMap<String, Object> mapData;
-}
