--- src/main/java/com/dottree/nonogrammers/domain/User.java
@@ -1,69 +1,0 @@
-package com.dottree.nonogrammers.domain;
-
-
-import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import org.hibernate.annotations.ColumnDefault;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@ToString
-@Table(name = "user")
-@EntityListeners(AuditingEntityListener.class)
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "userId")
-    private int id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String nickName;
-
-    @Column(columnDefinition = "text")
-    private String profileImgUrl;
-
-    @ColumnDefault("1")
-    private int statusCode;
-
-    @Column(nullable = false, unique = true)
-    private String baekjoonUserId;
-
-    @CreatedDate
-    @Column(nullable = false, updatable = false, columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
-    private LocalDateTime updatedAt;
-
-    @CreatedDate
-    @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP")
-    private LocalDateTime changedAt;
-
-    private String townName;
-
-    @Builder
-    public User(String email, String password, String nickName, String baekjoonUserId) {
-        this.email = email;
-        this.password = password;
-        this.nickName = nickName;
-        this.baekjoonUserId = baekjoonUserId;
-        this.profileImgUrl = "/images/default.png";
-        this.statusCode = 1;
-    }
-}
