+++ src/main/java/com/dottree/nonogrammers/domain/UserDTO.java
@@ -1,7 +1,5 @@
 package com.dottree.nonogrammers.domain;
 
-import com.dottree.nonogrammers.entity.User;
-import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDate;
