+++ src/main/java/com/dottree/nonogrammers/repository/UserRepository.java
@@ -2,7 +2,6 @@ package com.dottree.nonogrammers.repository;
 
 
 import com.dottree.nonogrammers.entity.User;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
